--- plan-service/src/main/java/com/nitb/planservice/PlanServiceApplication.java
@@ -1,15 +1,0 @@
-package com.nitb.planservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication(scanBasePackages = {"com.study.planservice", "com.study.common"})
-@EnableDiscoveryClient
-public class PlanServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PlanServiceApplication.class, args);
-    }
-
-}
