--- plan-service/src/main/java/com/nitb/planservice/entity/Plan.java
@@ -1,46 +1,0 @@
-package com.nitb.planservice.entity;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "plans")
-public class Plan {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "id", nullable = false)
-    private UUID id;
-
-    @Column(name = "creator_id", nullable = false)
-    private UUID creatorId;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "start_at", nullable = false)
-    private LocalDateTime startAt;
-
-    @Column(name = "end_at", nullable = false)
-    private LocalDateTime endAt;
-
-    @Column(name = "progress", nullable = false)
-    private Float progress;
-
-    @Column(name = "complete_at")
-    private LocalDateTime completeAt;
-
-    @Column(name = "team_id")
-    private UUID teamId;
-}
