+++ plan-service/src/main/java/com/study/planservice/PlanServiceApplication.java
@@ -1,0 +1,15 @@
+package com.study.planservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication(scanBasePackages = {"com.study.planservice", "com.study.common"})
+@EnableDiscoveryClient
+public class PlanServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PlanServiceApplication.class, args);
+    }
+
+}
