+++ plan-service/src/main/java/com/study/planservice/entity/PlanReminder.java
@@ -1,0 +1,30 @@
+package com.study.planservice.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "plan_reminders")
+public class PlanReminder {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @Column(name = "id", nullable = false)
+    private UUID id;
+
+    @Column(name = "plan_id", nullable = false)
+    private UUID planId;
+
+    @Column(name = "receiver_ids", nullable = false)
+    private String receiverIds;
+
+    @Column(name = "remind_at", nullable = false)
+    private LocalDateTime remindAt;
+}
