+++ plan-service/src/main/java/com/study/planservice/mapper/PlanReminderMapper.java
@@ -1,0 +1,28 @@
+package com.study.planservice.mapper;
+
+import com.study.planservice.entity.PlanReminder;
+import com.study.planservice.grpc.PlanReminderResponse;
+import com.study.planservice.grpc.PlanRemindersResponse;
+
+import java.util.List;
+
+public class PlanReminderMapper {
+    private PlanReminderMapper() {}
+
+    public static PlanReminderResponse toPlanReminderResponse(PlanReminder reminder) {
+        return PlanReminderResponse.newBuilder()
+                .setId(reminder.getId().toString())
+                .setRemindAt(reminder.getRemindAt().toString())
+                .build();
+    }
+
+    public static PlanRemindersResponse toPlanRemindersResponse(List<PlanReminder> reminders) {
+        List<PlanReminderResponse> responses = reminders.stream()
+                .map(PlanReminderMapper::toPlanReminderResponse)
+                .toList();
+
+        return PlanRemindersResponse.newBuilder()
+                .addAllReminder(responses)
+                .build();
+    }
+}
