+++ plan-service/src/main/resources/application-local.properties
@@ -1,0 +1,22 @@
+# H2 Database
+spring.datasource.url=jdbc:h2:mem:plandb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
+
+# SQL Init
+spring.sql.init.mode=always
+spring.sql.init.schema-locations=classpath:/database/schema.sql
+spring.sql.init.data-locations=classpath:/database/data.sql
+
+# JPA
+spring.jpa.hibernate.ddl-auto=none
+
+# Eureka client
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
+
+# Kafka
+spring.kafka.bootstrap-servers=localhost:9081
