+++ plan-service/src/main/resources/application-prod.properties
@@ -1,0 +1,15 @@
+# Postgres
+spring.datasource.url=jdbc:postgresql://${PLAN_SERVICE_DB_HOST:localhost}:${PLAN_SERVICE_DB_PORT:5432}/${PLAN_SERVICE_DB_NAME:plan_db}
+spring.datasource.username=${POSTGRES_USERNAME:user}
+spring.datasource.password=${POSTGRES_PASSWORD:123}
+
+# Hibernate
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.show-sql=true
+
+# Eureka client
+eureka.client.service-url.defaultZone=http://eureka-server:8761/eureka/
+
+# Kafka
+spring.kafka.bootstrap-servers=kafka:9080
