+++ plan-service/src/main/resources/application.properties
@@ -1,1 +1,21 @@
 spring.application.name=plan-service
+spring.profiles.active=${SPRING_PROFILES_ACTIVE:local}
+
+# Server
+grpc.server.port=9093
+server.port=8083
+
+# Eureka client
+eureka.client.register-with-eureka=true
+eureka.client.fetch-registry=true
+eureka.instance.prefer-ip-address=true
+
+# Kafka Producer
+spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
+
+# Kafka Consumer
+spring.kafka.consumer.group-id=plan-service-group
+spring.kafka.consumer.enable-auto-commit=false
+spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
