--- plan-service/src/test/java/com/nitb/planservice/PlanServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.nitb.planservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PlanServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
