+++ plan-service/src/test/java/com/study/planservice/PlanServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.study.planservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PlanServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
